Refactor move interfaces
