+++ client/src/components/Workflow/Invocation/Export/Plugins/DefaultFileExportPlugin.ts
@@ -1,8 +1,8 @@
-import { type InvocationExportPlugin } from "../model";
+import { type InvocationExportPlugin } from "@/components/Workflow/Invocation/Export/Plugins";
 
 export const DEFAULT_FILE_EXPORT_PLUGIN: InvocationExportPlugin = {
     id: "default-file",
-    title: "File",
+    title: "Compressed File",
     markdownDescription: `Export the invocation to a compressed File containing the invocation data in Galaxy native format.`,
     exportParams: {
         modelStoreFormat: "tgz",
